Always link system libraries
This changes the logic for linking system libraries from parsing the
input object and the `.symbols` files, to simply include the libraries
a link time in all cases.

The `.symbols` file logic were originally added as an optimization
would allow only the libraries needed to be considered at link time.

However, this optimization I believe was put in place for fastcomp
where the build time cost of including a library was significant
because all libraries would be passed to opt/llc.  The idea was that
minimal programs might be need to libc at all the there build
faster (even though the resulting output would not be effected due
to DCE).

It an optimization that apply with the wasm backend where linking is
fast and `.a` libraries are handled efficiently.  What is more there
are basically no real programs that don't use libc and for those
programs we already have the `-nostdlib` option.

The `.symbols` files are still used for some optional libraries such
as libhtml, libgl and libal.  However the plan is to also remove that
logic in STRICT mode too.   Once we do that we can avoid having to
run llvm-nm on each input object file.

See #8912.