Memory growth support for STANDALONE_WASM (#9588)
Remove the assertion on memory growth not being allowed in standalone mode.

Add emscripten_notify_memory_growth which is just called to notify. That will
show an error in wasm runtimes that don't support it.

Add a variation on libstandalone that calls that notification if growth is
enabled. That lets programs not using growth still be 100% pure.

Add a test for memory growth in "impure" standalone wasm (i.e., standalone mode
does not emit a 100% pure wasm yet).

Fix some bugs in emscripten_resize_heap and the preamble memory loading code,
that the test uncovered.