Remove try/catch cleanup handler from gen_struct_info.py
I was having trouble debugging failures in gen_struct_info.py and
notices it was aggressively cleanup up its temp files even in the failure
case.

This change means that if an exception is thrown for some reason the
temp files will be left around, allowing for debugging.  However on
success we still clean up.