Fix pthreads with closure compiler (#9569)
Turns out much of the pthreads runtime was not closure-safe. This PR fixes that,
using quoted strings as closure expects.

While doing that I noticed that some of the things in place for pthreads +
MODULARIZE would also help closure. In both cases the key thing is that
worker.js, that loads the main JS file, is separate from it, so it needs proper
interfaces and can't just rely on global variables (which in closure are
minified in the main JS, and in MODULARIZE are in another scope). So I removed
things like makeAsmExportAccessInPthread, makeAsmGlobalAccessInPthread,
makeAsmExportAndGlobalAssignTargetInPthread which makes the code simpler.

Because of that I was also able to remove a bunch of globals from worker.js
which further simplifies things, and should also help node.js workers (where the
 global scope is very different, #6567).

Most of this PR is straightforward according to the above notes. One slightly
tricky thing is the stack setup, which I refactored to use a new
Module.applyStackValues method that worker.js calls at the right time (as the
stack is set up after the JS loads).

Code size effects:

Before this PR: 69,567 bytes.
After this PR: 69,644.
After this PR, plus closure: 34,909.

So this adds only 77 bytes, while also making the code simpler + making it
possible to run closure and decrease size by 50%. In addition, I have followup
work to further reduce the non-closure code size too, so this regression will
be fixed (and more).