Add emsize.py tool
This is a rough emulation of the unix 'size' utility. It is meant to
include the JS glue file alongside the rest of the wasm sections
(which are supported by llvm-size already).

Currently this tool doesn't support any options; it's just enough
to be used by hand or by the LLVM testsuite runner script (which
outputs size info after building each test)