Optimize fmax/fmin/f etc. with wasm builtins (#9689)
The wasm builtins are very similar to the normal libc functions, except that
nans are handled differently. Keep the musl nan handling, and otherwise use the
builtins. This is ~41 bytes less in each fmaxf etc. function, and is 30% faster
on this silly benchmark:

#include <math.h>
#include <stdio.h>

int main() {
  union {
    int i;
    float f;
  } u;
  float sum = 0;
  const int N = 20000;
  for (int i = 0; i < N; i++) {
    for (int j = 0; j < N; j++) {
      u.i = ((i << 15) + j + 5085) ^ j ^ (i >> 2);
      sum += fmaxf(u.f, 0.5);
    }
  }
  printf("%.2f\n", sum);
}

After the speedup we are about equal with gcc natively.

Verified this does not change the output of our tests on this, and added more
test coverage, including of negative zero which libc is not guaranteed to get
right, but the implementation actually does, and using wasm builtins preserves
that.