Lazy-loading support (#9697)
This adds emscripten_lazy_load_code(), a function that when called will block
(using Asyncify) and load the complete program. This lets the initial download
not contain code that can be proven to be used only after those calls.

How this works is we emit the initial downloaded wasm after modifying it with
Binaryen to assume that those calls will not rewind, as we only unwind using
that binary (the optimizer can then remove a lot of code). We also emit the
later downloaded wasm, optimized the other way, to assume we only rewind but
never unwind. These transforms are done using ModAsyncify passes from Binaryen
WebAssembly/binaryen#2404

A new option ASYNCIFY_LAZY_LOAD_CODE is necessary to use this (as well as
ASYNCIFY itself).

The test shows an example where the initial download is less than half the size
of the later download. Note that in this model we do download some code twice
(the later download contains the code in the initial one), and we do add some
code size due to Asyncify support. So this will not be an obvious win for every
codebase. But if the later download contains unlikely code that may never be
used, and there is a lot of such code, and we can ignore indirect calls for
Asyncify's purposes, then the win can be significant.