Vary libgl on FULL_ES3 (#9762)
Fixes a regression from #9684 in which using WebGL2 + GetProcAddress would error
on missing FULL_ES3 methods.

If an app uses GetProcAddress, metadce can't remove GL functions (as it maps
string names to them, the user can ask for any one). And as a result we end up
with the entirety of libgl linked in even at the highest opt levels. And then
it's a link error as the JS library side intentionally does not link in these
functions (so that GetProcAddress will return "null" if asked for them, which
is the correct behavior if FULL_ES3 is not used).

Reduce the -O2 to -O1 in the relevant test, as it's really just testing for opts
vs not, which itself is of dubious importance; at least -O1 is faster to
compile.