Basic pthreads support for Node.js Workers (#9745)
This ended up tricky because of various node/web differences, like the global
scope, the event loop, APIs, etc. With help on #6567 I think this PR finally
manages to make it work, at least for an initial "hello world" type test.
Thanks to everyone on that issue and in particular @addaleax!

This moves some code from shell.js into side .js files, so that we can use it in
multiple places.

Fixes #6567 but as we add more tests I'm sure we'll find more issues, this just
shows basic functionality so far. We may not be able to easily reuse the
existing browser tests as they have browser-specific things in them, but we
can try.