Fix double-escaped paths on Windows
On Windows, paths are using \, which is escaped to \\ by llvm-dwarfdump in its output.

We need to unescape them before passing to source map encoder, as otherwise we end up with invalid paths and URLs that can't be resolved.

This wasn't caught before, because assertPathsIdentical was performing its own unescaping before comparing paths, which is not what happens in real source map consumers.

Fixes #9811.