Fail with ENOSYS on unsupported socketcalls (#9841)
There are some socketcalls that musl can call and are not handled
(for example socketpair). These currently abort the program instead of
returning an error code. Also, if the calling code doesn't use other
socket functions, there could be a linker error because of missing
ntohs, htons implementation. This dependency has been corrected.