Allow reusing slots in the WASM functions table.
We have a port of SQLite to WASM which calls addFunction for every SQL query
given that the results of the query are returned via a callback that is
different on every invocation. Given that removeFunction was a no-op, we would
quickly run out of slots in WASM table. This patch allows reusing table slots.