Use wrapper functions for growable heap views
Fixes pthreads + memory growth for some cases where the `HEAP*`
views were used with Atomics or other ways that didn't get
caught by the accessor transformations before.

For example `HEAP8.length` is transformed to use an accessor
as `GROWABLE_HEAP_I8().length` which updates the views if
necessary.

This should work for other library code which uses the `HEAP*`
views as well.

Fixes #9963