Remove SAFE_STACK option (#10069)
When we added it, we didn't realize that it does basically what fastcomp's
old STACK_OVERFLOW_CHECK=2 does, so we didn't need a new
option.

This makes STACK_OVERFLOW_CHECK=2 do what SAFE_STACK
did, which is check the stack at each increment. Note that that option
has been on by default in ASSERTIONS builds, so this PR has the
effect of making upstream behave the same as fastcomp, i.e, check
the stack in ASSERTIONS builds. A noticeable effect of this is in
the metadce tests.

The metadce tests had some windowsy \rs in them. I removed those
and added code to avoid adding them again.

Note that we can't use the renaming mechanism for aliases of
settings, as the values differ (1 vs 2). So I just removed the old
one, and it will show an error pointing to the new usage.