optimize Asyncify export instrumentation
Avoid 'arguments' usage, which causes a dynamic allocation and creates garbage
This is rather clunky and only implemented for functions that take 0, 1,
or 2 arguments, which is just enough for Fibers. Other functions still
use 'arguments'.

Also moved the fibers trampoline into Fibers and specialized it so that
the jump is static and can be inlined by closure.