Pthreads+closure EM_ASM fix (#10185)
Move EM_ASM reading logic from an always-emitted
readAsmConstArgs into a JS library function
emscripten_read_asm_const_args. We can then include
that function when needed, which is when EM_ASMs
exist, or when pthreads is enabled (as the proxy
logic uses it, even if EM_ASMs are not present; we
emit it without knowing about those).