Perform `npm install` after setting up emsdk
emsdk will run npm install in the version of emscripten that it
downloads, but we override with our own version from git and we need
to run `npm install` locally.

Previously we only did this for the socket tests which are currently
the only parts of emscripten that actually depend on `npm install`
having been run.  However we have plans to extend that.  See #9989.