Disable browser.test_cubegeom_normal_dap_far_glda_quad on firefox (#10290)
I investigated that new consistent CI failure, and it works
locally for me on firefox dev, so it's possible there
is some weird combination of GPU driver on the bots +
a recent firefox change that interacts badly. Hard to know
where to send a bug report. Anyhow, this is a minor test so
we shouldn't be losing much coverage here (the test doesn't
really stress anything special, just yet another GL operation).