Add a closure suppression only for fastcomp asm.js code (#10333)
It is not necessary for wasm2js with the wasm backend,
where things look very differently. This code just tried to
parse the wasm2js code and ended up with bad results,
putting the suppression in the middle of irrelevant code,
which was ignored by closure as a bad annotation, but
if warnings are errors then it is noticeable.