[docs] Mention that Mozilla is removing IRC, and that there is a Discord room (#10165)
Currently the Discord room is where most of us are
congregating, and most of the wasm community in
general. However, as disussed on github and the
mailing list, we are open to other places too, if someone
wants to investigate options there - this PR is basically
just updating on the Mozilla removal of IRC, and that
Discord is already used in practice.