Basic DWARF testing (#10298)
The test ensures each interesting line is mapped to one
debug line, and verifies that points to a proper place in
the wasm binary. (We track 3 calls that the optimizer
can't reorder or inline, so we can figure out where that is.)
That is, the test matches up binary locations in the final
wasm with the DWARF debug line info and with the source
line it refers to, seeing all three are in sync.