Allow string values for PTHREAD_POOL_SIZE (#10263)
This makes PTHREAD_POOL_SIZE a string, which allows stuff
like -s PTHREAD_POOL_SIZE=navigator.hardwareConcurrency
in addition to the old -s PTHREAD_POOL_SIZE=8 etc. (This is
fully backwards compatible with the old notation only because
our parser is smart enough to not require unnecessary
quotes for string values.) The value the user provided is simply
emitted in the JS, and used, so it could be anything that is valid
JS.

For example, it can now be set to

  navigator.hardwareConcurrency

which would automatically create a number of Workers equal
to the cores in the machine as reported by the browser. If
the compiled application then tries to make a threadpool of
one per core, the number of Workers will be exactly suitable.