Enable reporting of undefined symbols in wasm-ld (#10350)
When enabled, wasm-ld will report more detailed information about the
undefined symbols such as which object file/funcion required the
symbols.

This works by generating the list of all possible JS-defined symbols
and passing that list of wasm-ld so it can determine which symbols
are permitted to be undefined.

The cost of this feature is that we run the js compiler twice, once
before wasm-ld and once after the link once wasm-emscripten-finalize
has been run.  On my machine this takes around 700 so I'm not enabling
this by default yet.