Fix NO_EXIT_RUNTIME + USE_PTHREADS + MODULARIZE (#10158)
When NO_EXIT_RUNTIME the local noExitRuntime must
be set. In pthreads mode, the workers load the main JS
using importScripts which puts the main JS's noExitRuntime
into worker.js's global scope, so the local has the right
value. However, in modularize mode the imported script
has its contents inside a function scope. So we were not
seeing the proper value of noExitRuntime.

This makes us call a helper function, exported from the
module, to access the value of noExitRuntime.