Adjust to Closure compiler convention that parseInt() must always be called with two input parameters.
Closure compiler has their own convention that the second radix argument to parseInt() is not optional, but then unfortunately Closure compiler also does not optimize out the second argument when one does pass it.

This PR adds two new functions jstoi_s and jstoi_q, similar to spirit with the C "atoi" function.

jstoi_s is not actually used, but I think it's good to have for completeness.