Remove old files from musl source tree
When musl was updated to version 1.1.5 in #4813 it looks like files
that were removed upstream were not removed locally.

Because we recursively build some directories this means we were
build old files. For example the src/locale/is*.c family of functions
were removed from musl in d89fdec51b5849ebdf8000ff1c2fb49878004f39
but remained in our tree after the update.

This fixes some strange failures we've been seeing on the CI builders
where the name of a symbol alternates between __isxdigit_l (the new
name) and isxdigit_l (the old name).   This is due to the fact that
this symbol was defined in two different object files and filesystem
ordering was determining which object was included first by the linker.

This change was automatically generated by checking out musl at v1.1.15
and running:

`for f in $(find . -type f); do if [[ ! -e ../../../../../musl/$f ]]; then echo $f; fi; done  | grep -v emscripten | xargs git rm`