Remove unused EM_PROXIED_SYSCALL special hanling of syscalls (#10477)
As far as I can tell syscalls currently get proxied just like any other
JS function.  I couldn't see this functionality used anywhere in the
codebase.  I'm in the process of updating the syscalls that is not
compatible with this mechanism in that it removes the first argument
(which was redundant) and also no longer uses varargs.