Fix minifyGlobals on duplicate definitions (#10522)
In JS it is valid to write

var a;
var a;

but that wasn't handled properly in minifyGlobals. This only
became a problem now (years after the pass was written)
because we added a loop in the mem init that uses i, and if
there are enough imports one of them might be i as well,
leading to incorrect minification and a bugreport from the
mailing list ("js output not running with 1.39.8").