fastcomp: error on unrecognised linker inputs
Also, make the error messages consistent with wasm-ld when it
sees invalid input files.

Also, simplify the running wasm-opt and allow its error code to be
handled by `check_run` and allows its errors/warnings to be displayed
on stderr directly.

This change was split of from a larger change relating to handling
of linker inputs: #10560