sysconf() support in standalone mode (#10535)
This doesn't replace the JS version since from JS we can
check the number of CPUs and other useful things.

This makes the C version return the same values as JS,
which are probably pretty arbitrary, but seems like no
reason to cause any churn for users. However one value
here (_SC_THREAD_PRIORITY_SCHEDULING) was 0,
which it turns out is simply invalid (values must be positive
or negative), which I fixed.