Fix some failing LTO tests (#10562)
All but one of the non-slow tests now pass under LTO:

  EMTEST_SKIP_SLOW=1 ./tests/runner.py wasmlto2

Remaining failure is test_floatvars.

This changed moved a few source files from libc to libc_rt_wasm (the
parts of libc can are excluded from LTO because they contain libcall
implementations).  This means there was a slight impact on the output
due to the `-O2` vs `-Os`.