-gside option: emit DWARF in a wasm on the side (#10568)
When used, the main wasm file has no dwarf in it, but a
.debug.wasm file on the side does. This is similar to
split-dwarf in gcc/clang, but simpler. There is also a
custom section in the wasm that refers to the debug
file. We may evolve this further but for now it should be
useful.

Implemented internally using a `SIDE_DEBUG` flag.