Fix parsing of TOTAL_MEMORY=64MB (#10581)
This is a regression from #10566 which renamed that
setting to INITIAL_MEMORY, and #10449 which added
a test using that setting in that way, and landed close
enough that we didn't notice it was doing so. It should
have worked since we tried to support the old name of
the setting, but this uncovered a bug, so it's actually
good that it broke!

The bug is that we have special handling for expanding
things like 64MB into a number, and that handling did
not work on aliases. As a solution, map aliases very
early in parsing.