Move musl internal lock functions from JS to C
__lockfile/__unlockfile were already patch to support __EMSCRIPTEN__
in system/lib/libc/musl/src/stdio/__lockfile.c so just start compiling
that file.

__lock/__unlock were implemented as no-ops in two different places
so move them to extras.c for now.

The fact that both these pairs of functions are currently nop'd out
is almost cetainly a problem for pthreads builds, but not one that
is solved by this PR :)

To make this work I also had to move extras.c back inside libc
but keep extras_fastcomp.c in its own bc file for the reasons stated
therein.