Refactor WarningManager. NFC.
Use the actual command line warnings names as the settings keys to make
it easier to grep for usage of, say `separate-asm` in the codebase.
This simplifies the code too.

Force all call sites to pass thier own message.  There was no use of the
shared messages within the codebase and having the error messages live
near the error site makes the code easier to navigate.

Both this changes were made in preparation for adding a new controllable
warning regarding the use of legacy settings.