memset() refactoring (#10656)
I noticed that in some builds we'd end up with memset in JS, instead
of in compiled code. That's because we have memset implemented in
our JS library code. That was useful in the old days before compiler_rt
but today it's risky - it can lead to surprising slowness and bloat. It's
also an old hack for fastcomp that we'll want to remove anyhow in
time.

This ifdefs out that memset in the JS library for upstream. I didn't try
to also fix fastcomp as linking is different there, and it's not worth the
effort.

This removes memset from DEFAULT_LIBRARY_FUNCS_TO_INCLUDE
which is an old hack. This removal is necessary as otherwise we look
for it in the JS library, where it isn't any more, and error. (In fastcomp,
we add it back in manually.)

There is a potentially noticeable difference here: in the past you could
do x__deps: ['memset'] in a JS library, that is, depend on the JS
version of memset. We have to disallow that if we want to always use
the compiled optimal version. The fix for JS library code that wants to
use memset is the usual: set it up in deps_info.json, that is, tell the
linker that if x will be used from the JS library, then memset is needed
from the system libraries. Or, without modifying deps_info.json - say,
in a user's JS library - adding _memset to EXPORTED_FUNCTIONS
will work. This PR adds a warning to help figure that out.

The one slightly annoying thing in this PR is that we need to add code
in system_libs.py to handle the sanitizers specifically: sadly the
usual deps_info mechanism does not work since we scan only user files
for things, and not libraries (to be able to scan libraries, we'd need to
somehow figure out which of their object files will actually be linked in -
but only lld knows that).