Fix a wasm2js minification bug with localNames being for the whole set of functions (#10682)
This is an embarrassing bug: a small typo caused us to scan the whole
set of functions being optimized (we optimize in chunks) for locals, as
opposed to just the actual locals in the current function. As a result, we
thought each function had a lot more locals. That prevented us from
using smaller local names in the minification.

This improves wasm2js code sizes by 3% on a few quick benchmarks
I did, and on the minimal runtime tests. It also speeds up compilation by
a factor of 2 (!) on a quick benchmark. These two benefits may be
even larger on a huge codebase, so it makes me wonder (cc @juj who
mentioned wasm2js code size issues).

May fix both #10673 and #10675, as this is more than just a nice
improvement, it also could have been buggy before: consider if we
need to call a function foo but it looks like foo is one of our locals,
then we'd get confused and call something wrong.

This was exceptionally hard to debug, as it only happens on big-enough
projects (where there are enough locals and globals to collide), namely
those two bugs just linked. Also one of them can't provide source and the
other only reproduces it on CI but not locally, which appears to be due
to nondeterminism - part of which is due to this very issue, but I have
another PR in the works as well. Also making this hard to debug was
that the problem is in name minification, so preserving function names -
the most minimal amount of debug info - hides the bug...