Debug Info: Remove -gforce_dwarf, use -g normally like a native compiler (#10509)
This fixes #10325. It removes the temporary -gforce_dwarf flag which
emitted DWARF, and instead makes -g emit DWARF like a normal
native compiler. This is useful as DWARF support is improving and
we are starting to test it.

This may be a user-noticeable change:

-g on object files leads to full DWARF generation, normally. That
may be larger than before, where we only emitted debug lines
since we assumed DWARF didn't work anyhow.

-g at link time leaves DWARF info and updates it, which again,
can be larger than a build with -g was before.

The size change should only be in -g builds, so not production ones,
and the behavior matches native compilers, so hopefully this won't
be too disruptive. But it may slow down builds. For example if
someone builds object files with -g, then links without -g, the
object files will be larger and linking take longer, so such
unnecessary -g usage may be annoying in a way it wasn't before
(but again, that's the same natively too).

Also refactor away the opt_level and shrink_level flags in emcc.py,
which duplicated OPT_LEVEL / SHRINK_LEVEL in the Settings
object, and which caused a bug for me here.

Also simplify test_lua to remove some old fastcomp code that
made it hard to debug why it was failing here, and isn't needed.

This PR uncovered
https://bugs.llvm.org/show_bug.cgi?id=44920
which blocked landing, but has been fixed.