Optimize glUniformMatrix4fv for size and also pthreads+growth by hoisting the heap. (#10689)
This ends up with a single call to GROWABLE_HEAP_F32() instead
of 16. In theory the JS engine can inline all those, but this is an extreme
case with 16 accesses in a row, and best to be sure of good results.

It also makes the code a little smaller (also without pthreads+growth).