Properly handle metadce of MINIMAL_RUNTIME exports (#10663)
We handled imports, but not exports. As a result, some things in
the wasm binary were not cleaned up by metadce. I noticed this
when comparing builds with minimal and normal runtime - the JS
was much smaller in minimal, but the binaries were a little bigger
due to stuff like unnecessary dynCalls, things like stackAlloc, etc.

This adds a metadce test for minimal runtime.

After this, minimal runtime has similar wasm sizes to the normal
runtime.