Sort functions properly in wasm2js (#10685)
The code first splits up by function, then sorts. But the splitting
code only worked if newlines were present. This fixes that so that
we split properly even without newlines.

This is not compatible with fastcomp, which has source maps
support that depends on specific line numbers, and we should
not split without newlines (that would add a newline). So this
forks that code, which is a little unfortunate, but we'll remove it
with fastcomp.

This makes wasm2js almost entirely deterministic, but there are
newline differences that we can look into separately.