Fix browser.test_offset_converter on CI (#10722)
This random failure has annoyed us for a while, and I think
this PR fixes it. The underlying problem was that the
JS offset converter misparsed memory and table limits.

As a result of that misparsing it was reading random memory
in the wasm for any imports after the memory. There was
only one such import, the table, which we don't care about,
and that is why this went unnoticed for so long. We got a
garbage value for the kind of that import, and just fell
through in the switch and did nothing (this PR adds an
assert to prevent that going forward), but we didn't need
to do anything for the table so the problem wasn't obvious.

The reason this would sometimes fail on CI is that the testcase
has an assertion in the C code. The assertion ends up
embedding the full path to the source file, and that path
is of course different between CI and when you run locally,
and on CI it will also vary randomly. That path happens to
be where that garbage data was read. Rarely, it happened
to have a 0 in the "right" place, which made us think that
the last import's type was a function. It then added 1 to the
number of function imports, which led to an off-by-one
error in looking up function names.