Fix browser test harness on python3.7 (#10738)
Ensure we are at a valid current dir at all times.

I'm not sure why I only saw this locally, as on CI it works fine.
The symptom was that after the first test finishes, we'd get a
ping, fail to respond as the python http server would
error "is not a directory", and the ping would take that as
no more tests to be run and shut down. This was workaroundable
locally with EMTEST_SAVE_DIR=1 which avoids deleting
the temp dir.