Add `make dist` target for creating a distibutable version of emscripten
This target creates an archive that excludes parts of the source tree
that are not needed by end users.

Significantly we don't include `.git` or `tests`.

The resulting archive is 64Mb compressed and 77Mb uncompressed where as
the unfiltered contents currently in emsdk is 114Mb compressed and nd
131Mb uncompressed.

Currently the only use for the rules is to upload an artifact to github
CI.  However as a followup we can start using this rule when building
the emsdk archives to save significant amount data for our users.