Add internal CAN_ADDRESS_2GB flag which is set for 2GB+ heap support (#10601)
To support that mode add a JS optimizer pass that fixes up
pointers in JS libraries. Doing this as a pass is nice because
our JS libraries can still be written with signed pointers,
which is smaller and may be faster on some VMs (signed
values fit in 31-bit smis). This will fix up user code as well,
automatically.

The pass looks for HEAP8 etc. symbols on which we do
HEAP8[something] and ensures the pointer is unsigned. It
also fixes up .set() and other calls that receive pointers.

One compromise the pass makes is that it looks not just
for the canonical HEAP8 etc. names but also HEAP and
heap. Those are used in our JS library code a bunch, so
fixing those up too prevents a lot more work in that code, and
can help user code too. There is some risk to doing so, if
a user has code with that name but actually doesn't want
the unsigning; however, the risk is small (we just force their
indexes on an array to be non-negative; likely they don't
want that anyhow!), and I think the reverse risk (a user has
code that should be changed, but forgets) is much more
likely, so this seems safest. A later PR will add our JS library
fixes, which will show that even with this extension there is
still code that must be handled manually, so I think it's best
to do as much automatically as we can.

This also changes MAXIMUM_MEMORY to be 2GB. That
is effectively no change from the previous -1 (no limit)
since 2GB was the absolute limit anyhow. This has two benefits:
first, people must opt-in explicitly to getting 2GB+ heap support
with the code size downside, and also we expect only some
browsers will support 2GB+ heaps for now, so requiring explicit
opt-in from devs is safer to avoid web compat issues (i.e., only
users that really need such heaps will set them, and then they
may only work in some browsers, but at least that won't happen
for users that could actually run in all browsers).

Note that emmalloc is not designed to work with 2GB+ heaps,
so show a useful error in that case. This is a minor breaking
change for users in that building with emmalloc + memory growth
will hit the error, forcing the user to set a maximum to the growth.

This is part of a set of 3+ PRs for #6566, together with #10600