Add EMIT_EMSCRIPTEN_LICENSE option to emit license info in JS (#10770)
As the person that created this project, I've never thought it's necessary
to add license info in the output JS, since it's compiler output. However,
IANAL. This adds a compiler flag so people that do want to emit the
license info can do so.

Also move all EMIT_ flags into a single area in the settings file.

For simplicity the emitted license only mentions MIT. I gave Emscripten
two license options back in the day, MIT and LLVM, the latter just to
avoid issues with sharing code with LLVM. That ended up not necessary.
The LLVM license is technically still there, but not worth mentioning in a
SPDX-License-Identifier because I hear automatic tools may fail
to parse a dual license (if anyone does want to use the defunct LLVM
license instead of MIT, please let me know).

This is not supported in fastcomp since the JS optimizer pipeline is
more complex there, and not worth the work at this time given fastcomp
is deprecated.