Use `emscripten-revision.txt` file to store extra version information
Rather than adding this to the existing `emscripten-version.txt` file.
Just in case anyone was parsing this outside of emscripten itself.

This is followup on #10789 based on feedback in that PR.

Also use `git rev-parse HEAD` to get the git revision for git users
which is slightly more efficient/reliable that trying to parse
`git show` output.