Reset selector before personality function call
In libunwind, we need to reset the selector every time before we call
the personality function. If we don't do that, if the selector was set
to a specific value in the previous call and there's no matching `catch`
clause in the current call so the selector is not set, the previous
value will still remain and cause incorrect execution.

`test_exceptions_typed` should run fine after this and
https://github.com/llvm/llvm-project/commit/c87b5e7e22b2df92021ac5fcc69160901a5841a9.