Remove wasm_simd128.h
This header has moved to the LLVM project, where it can be more widely
used. This PR removes the now-redundant contents of the wasm_simd128.h
header in this repository and replaces them with a redirection to the
version that now lives in the clang distribution. However, the primary
tests for the header are still in this repository. In the process of
moving the header to clang, its handling of `-funsigned-char` was
corrected, so this PR also adds testing with that flag.