Remove dummy/fake version of malloc/free
These symbols should always be part of the native code.  We
unconditioanlly add them to EXPORTED_FUNCTIONS so they should always
exist in the output binary.

This change help to unblock #10826.  The logic in #10826 assumes that
JS library function don't also appear in EXPORTED_FUNCTIONS.  i.e. that
a symbol can either be an exported function *or* a JS library function
but not both.