Add IGNORE_MISSING_MAIN setting and --no-entry argument (#10827)
This settings, when disabled, forces the developer to be explicit
about the presence or absence of `main` their program.  It is enabled
by default but disabled in STRICT mode.

In when disabled, if you want to build a program without a `main` 
function you must either build with `--no-entry` or pass a list of
EXPORTED_FUNCTIONS the doesn't include `_main`.

The reason for adding the extra `--no-entry` option is that this
is what `wasm-ld` recommends in its corresponding error message.  Its
also a lot easier to type than `-s EXPORTED_FUNCTIONS=[]`.

This change is a step towards enabling better error reporting from lld
by enabling the removal of the `--allow-undefined` option.