Use malloc() in GL context creation only with pthreads (#10813)
A GL context is an opaque integer from the user's perspective. Internally
it is a heap allocation, and in pthreads we use that location in memory
to coordinate GL things between threads. However, in a non-pthreads
build that isn't necessary - we can just use an integer ID. Doing so
avoids using malloc/free, which in small programs can avoid bringing in
malloc/free unnecessarily.

This helps several of the code size tests by around 3% (see diff).